Fix 32-bit offset overflow in Metal sort kernel
Cast tid.y to size_t before multiplying with stride to avoid
overflow when segment_index * stride exceeds 2^31.
For shapes like (65537, 65536), row 65536 has offset 2^32 which
overflows int32 and causes incorrect sort results.

Fixes #4538

+++ mlx/backend/metal/kernels/sort.h
@@ -280,8 +280,8 @@ struct KernelMergeSort {
       uint3 tid [[threadgroup_position_in_grid]],
       uint3 lid [[thread_position_in_threadgroup]]) {
     // tid.y tells us the segment index
-    inp += tid.y * in_stride_segment_axis;
-    out += tid.y * out_stride_segment_axis;
+    inp += static_cast<size_t>(tid.y) * in_stride_segment_axis;
+    out += static_cast<size_t>(tid.y) * out_stride_segment_axis;
 
     // Copy into threadgroup memory
     for (short i = lid.x; i < N_PER_BLOCK; i += BLOCK_THREADS) {
